# openapi.yaml
openapi: 3.0.3
info:
  title: Advertising Platform API
  description: |
    Domain-driven API for managing advertising campaigns, tracking performance,
    and optimizing conversions. Built from bounded contexts rather than raw SQL schema.
    ## Key Features
    - **Domain-Driven Design**: API designed around business domains, not database tables
    - **Property-Based Testing Ready**: Comprehensive OpenAPI spec for automated testing
    - **Multi-Level Tracking**: 5-level sub-tracking (sub1-sub5) and affiliate tracking
    - **Fraud Detection**: Built-in bot detection and fraud scoring
    - **A/B Testing**: Landing page optimization with control/test groups
    - **Real-time Analytics**: Performance metrics with multiple breakdowns
  version: 1.0.0
  contact:
    name: API Support
    email: api@advertising-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:5000/v1
    description: Local development server (port 5000)
  - url: http://127.0.0.1:8001/v1
    description: Local development server (port 8001)
  - url: https://api.advertising-platform.com/v1
    description: Production server
  - url: https://staging-api.advertising-platform.com/v1
    description: Staging server
# Global rate limiting (applied to all endpoints)
...
security:
  - bearerAuth: [ ]
  - apiKey: [ ]
# ============================================================================
# SECURITY SCHEMES
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated users
      x-bearerInfoFunc: mock_server.bearer_token_auth
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for partner integrations
    oauth2:
      type: oauth2
      description: OAuth 2.0 authorization for third-party applications
      flows:
        authorizationCode:
          authorizationUrl: https://auth.advertising-platform.com/oauth/authorize
          tokenUrl: https://auth.advertising-platform.com/oauth/token
          scopes:
            # Campaign scopes
            campaign:read: Read campaign data
            campaign:write: Create and modify campaigns
            campaign:delete: Delete campaigns
            # Landing page scopes (NEW)
            landing:read: Read landing pages
            landing:write: Create and modify landing pages
            # Offer scopes (NEW)
            offers:read: Read offers
            offers:write: Create and modify offers
            # Analytics and traffic
            analytics:read: Read analytics and performance data
            traffic:read: Read traffic and click data
            # Admin
            admin: Full administrative access
  # ============================================================================
  # COMMON SCHEMAS
  # ============================================================================
  schemas:
...
    Error:
      type: object
      description: Standardized error response format
      required:
        - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: VALIDATION_ERROR
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - INVARIANT_VIOLATION
                - UNAUTHENTICATED
                - UNAUTHORIZED
                - METHOD_NOT_ALLOWED
                - NOT_ACCEPTABLE
                - UNSUPPORTED_MEDIA_TYPE
                - TOO_MANY_REQUESTS
                - INTERNAL_SERVER_ERROR
                - SERVICE_UNAVAILABLE
            message:
              type: string
              description: Human-readable error message
              example: "Validation failed: name is required"
            details:
              type: object
              description: Optional object for detailed validation errors
      example:
        error:
          code: "VALIDATION_ERROR"
          message: "Validation failed: name is required"
          details:
            field: "name"
            issue: "required"
    # Enums
...
paths:
...
  /forms/submit:
    post:
      summary: Submit form data
      description: Handle form submissions and lead capture from landing pages
      operationId: submitForm
      tags: [ Forms ]
      security: [ ]  # Public endpoint for form submissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                form_id:
                  type: string
                  example: "contact_form"
                campaign_id:
                  type: integer
                  example: 123
                fields:
                  type: object
                  additionalProperties: true
                  example:
                    name: "John Doe"
                    email: "john@example.com"
                    phone: "+1234567890"
                metadata:
                  type: object
                  additionalProperties: true
                  example:
                    user_agent: "Mozilla/5.0..."
                    referrer: "https://facebook.com"
      responses:
        '200':
          description: Form submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  lead_id:
                    type: string
                    example: "lead_123"
                  message:
                    type: string
                    example: "Form submitted successfully"
        '400':
          description: Invalid form data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  # ============================================================================
  # RETENTION CAMPAIGNS
  # ============================================================================
...